fix: split components into different files
