+++ packages/tradershub/src/components/ContentSwitcher/ContentTabPanel.tsx
@@ -1,0 +1,10 @@
+import React, { PropsWithChildren } from 'react';
+
+/**
+ * Panel for displaying content.
+ * @param {PropsWithChildren} props The props for the component.
+ * @returns {JSX.Element} The rendered component.
+ */
+const ContentTabPanel = ({ children }: PropsWithChildren) => <>{children}</>;
+
+export default ContentTabPanel;
